# crates/gateway/openapi/admin-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "gateway",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.0.0"
  },
  "paths": {
    "/api/v1/admin/api-keys": {
      "get": {
        "tags": [
          "crate::http::api_keys"
        ],
        "operationId": "list_api_keys",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AdminApiKeysPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "crate::http::api_keys"
        ],
        "operationId": "create_api_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_CreateApiKeyResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/admin/api-keys/{api_key_id}": {
      "patch": {
        "tags": [
          "crate::http::api_keys"
        ],
        "operationId": "update_api_key",
        "parameters": [
          {
            "name": "api_key_id",
            "in": "path",
            "description": "API key identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_UpdateApiKeyResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/admin/api-keys/{api_key_id}/revoke": {
      "post": {
        "tags": [
          "crate::http::api_keys"
        ],
        "operationId": "revoke_api_key",
        "parameters": [
          {
            "name": "api_key_id",
            "in": "path",
            "description": "API key identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RevokeApiKeyResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminApiKeyModelView": {
        "type": "object",
        "required": [
          "id",
          "key",
          "tags"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AdminApiKeyServiceAccountOwnerView": {
        "type": "object",
        "required": [
          "id",
          "name",
          "key",
          "team_id",
          "team_key",
          "team_name"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "team_id": {
            "type": "string"
          },
          "team_key": {
            "type": "string"
          },
          "team_name": {
            "type": "string"
          }
        }
      },
      "AdminApiKeyUserOwnerView": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "AdminApiKeyView": {
        "type": "object",
        "required": [
          "id",
          "name",
          "prefix",
          "status",
          "owner_kind",
          "owner_id",
          "owner_name",
          "model_keys",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "last_used_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "model_keys": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "owner_email": {
            "type": [
              "string",
              "null"
            ]
          },
          "owner_id": {
            "type": "string"
          },
          "owner_kind": {
            "type": "string"
          },
          "owner_name": {
            "type": "string"
          },
          "owner_service_account_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "owner_service_account_team_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "owner_service_account_team_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "owner_team_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "prefix": {
            "type": "string"
          },
          "revoked_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "CreateApiKeyRequest": {
        "type": "object",
        "required": [
          "name",
          "owner_kind",
          "model_keys"
        ],
        "properties": {
          "model_keys": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "owner_kind": {
            "type": "string"
          },
          "owner_service_account_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "owner_team_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "owner_user_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Envelope_AdminApiKeysPayload": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "items",
              "users",
              "service_accounts",
              "models"
            ],
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AdminApiKeyView"
                }
              },
              "models": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AdminApiKeyModelView"
                }
              },
              "service_accounts": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AdminApiKeyServiceAccountOwnerView"
                }
              },
              "users": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AdminApiKeyUserOwnerView"
                }
              }
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        }
      },
...
      "Envelope_CreateApiKeyResponse": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "api_key",
              "raw_key"
            ],
            "properties": {
              "api_key": {
                "$ref": "#/components/schemas/AdminApiKeyView"
              },
              "raw_key": {
                "type": "string"
              }
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        }
      },
...
      "Envelope_RevokeApiKeyResponse": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "api_key"
            ],
            "properties": {
              "api_key": {
                "$ref": "#/components/schemas/AdminApiKeyView"
              }
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        }
      },
...
      "Envelope_UpdateApiKeyResponse": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "api_key"
            ],
            "properties": {
              "api_key": {
                "$ref": "#/components/schemas/AdminApiKeyView"
              }
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        }
      },
...
      "ResponseMeta": {
        "type": "object",
        "required": [
          "generated_at"
        ],
        "properties": {
          "generated_at": {
            "type": "string"
          }
        }
      },
...
      "UpdateApiKeyRequest": {
        "type": "object",
        "required": [
          "model_keys"
        ],
        "properties": {
          "model_keys": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ogw_session"
      }
    }
  }
}